random uint32 price

+++ cmd/photon/provider.go
@@ -187,8 +187,8 @@ func doProviderRunCommand(ctx context.Context, cmd *cobra.Command, args []string
 				return
 			}
 
-			// randomize price [0, price]
-			price = rand.Int31n(price + 1)
+			// randomize price
+			price = uint32(rand.Int31n(int32(price) + 1))
 
 			ordertx := &types.TxCreateFulfillment{
 				Fulfillment: &types.Fulfillment{
@@ -229,9 +229,9 @@ func doProviderRunCommand(ctx context.Context, cmd *cobra.Command, args []string
 	return common.MonitorMarketplace(ctx.Log(), ctx.Client(), handler)
 }
 
-func getPrice(ctx context.Context, addr base.Bytes, seq uint64) (int32, error) {
+func getPrice(ctx context.Context, addr base.Bytes, seq uint64) (uint32, error) {
 	// get deployment group
-	price := int32(0)
+	price := uint32(0)
 	path := addr.EncodeString() + string(seq)
 	group := new(types.DeploymentGroup)
 	result, err := query.Query(ctx, path)
